pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "claw"
version = "0.1.0"
description = "The Claw — Local-first voice AI agent with MCP tool calling"
requires-python = ">=3.12"
dependencies = [
    "sounddevice>=0.5",
    "numpy>=1.26",
    "openwakeword>=0.6",
    "onnxruntime>=1.17",
    "faster-whisper>=1.0",
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
    "openai>=1.50",
    "mcp>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sse-starlette>=2.0",
    "jinja2>=3.1",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "pyyaml>=6.0",
    "watchfiles>=0.24",
    "python-dotenv>=1.0",
    "python-dateutil>=2.9",
    "httpx>=0.27",
    "segno>=1.6",
]

[project.optional-dependencies]
youtube-music = [
    "ytmusicapi>=1.7",
    "yt-dlp>=2024.1",
]
tts = [
    "piper-tts>=1.2",
]
google = [
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-auth>=2.0",
]
gemini = [
    "google-generativeai>=0.8",
    "cryptography>=42.0",
]
all = [
    "claw[tts,youtube-music,google,gemini]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
claw = "claw.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
